+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/TigerConfigurationKeyString.java
@@ -6,6 +6,7 @@ package de.gematik.test.tiger.common.config;
 
 import java.util.List;
 import lombok.Data;
+import org.apache.commons.lang3.StringUtils;
 
 @Data
 public class TigerConfigurationKeyString {
@@ -16,8 +17,8 @@ public class TigerConfigurationKeyString {
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
-    if (o instanceof String) {
-      return value.equalsIgnoreCase((String) o);
+    if (o instanceof String s) {
+      return value.equalsIgnoreCase(s);
     }
     if (!(o instanceof TigerConfigurationKeyString)) return false;
 
@@ -44,5 +45,9 @@ public class TigerConfigurationKeyString {
 
   public String asString() {
     return value;
+  }
+
+  public boolean isNotEmptyKey() {
+    return StringUtils.isNotEmpty(value);
   }
 }