+++ tiger-rbel/src/main/java/de/gematik/rbellogger/writer/tree/RbelContentTreeNode.java
@@ -4,11 +4,10 @@
 
 package de.gematik.rbellogger.writer.tree;
 
-import de.gematik.rbellogger.RbelContent;
 import de.gematik.rbellogger.RbelLogger;
 import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.rbellogger.data.RbelMultiMap;
-import de.gematik.rbellogger.data.facet.RbelFacet;
+import de.gematik.rbellogger.util.RbelPathAble;
 import de.gematik.rbellogger.util.RbelPathExecutor;
 import de.gematik.rbellogger.writer.*;
 import de.gematik.test.tiger.common.jexl.TigerJexlContext;
@@ -21,7 +20,7 @@ import javax.annotation.Nullable;
 import lombok.Getter;
 import lombok.Setter;
 
-public class RbelContentTreeNode implements RbelContent {
+public class RbelContentTreeNode extends RbelPathAble {
 
   private RbelMultiMap<RbelContentTreeNode> childNodes;
   private Map<String, String> attributeMap = new HashMap<>();
@@ -115,10 +114,11 @@ public class RbelContentTreeNode implements RbelContent {
   }
 
   @Override
-  public Optional<RbelContentTreeNode> getFirst(String key) {
+  public Optional<RbelPathAble> getFirst(String key) {
     return getChildNodesWithKey().stream()
         .filter(entry -> entry.getKey().equals(key))
         .map(Map.Entry::getValue)
+        .map(RbelPathAble.class::cast)
         .findFirst();
   }
 
@@ -140,7 +140,6 @@ public class RbelContentTreeNode implements RbelContent {
     return childNodes.getValues().stream().collect(RbelMultiMap.COLLECTOR);
   }
 
-  @Override
   public Optional<RbelContentTreeNode> findElement(String rbelPath) {
     final List<RbelContentTreeNode> resultList = findRbelPathMembers(rbelPath);
     if (resultList.isEmpty()) {
@@ -157,7 +156,6 @@ public class RbelContentTreeNode implements RbelContent {
             + " elements, expected only one!");
   }
 
-  @Override
   public RbelContentTreeNode findRootElement() {
     RbelContentTreeNode result = this;
     RbelContentTreeNode newResult = result.getParentNode();
@@ -168,7 +166,6 @@ public class RbelContentTreeNode implements RbelContent {
     return result;
   }
 
-  @Override
   public String findNodePath() {
     LinkedList<Optional<String>> keyList = new LinkedList<>();
     final AtomicReference<RbelContentTreeNode> ptr = new AtomicReference<>(this);
@@ -188,11 +185,9 @@ public class RbelContentTreeNode implements RbelContent {
 
   @Override
   public List<RbelContentTreeNode> findRbelPathMembers(String rbelPath) {
-    return new RbelPathExecutor(this, rbelPath)
-        .execute(RbelContentTreeNode.class).stream().map(RbelContentTreeNode.class::cast).toList();
+    return new RbelPathExecutor<>(this, rbelPath).execute();
   }
 
-  @Override
   public Optional<String> findKeyInParentElement() {
     return Optional.of(this)
         .map(RbelContentTreeNode::getParentNode)
@@ -214,25 +209,10 @@ public class RbelContentTreeNode implements RbelContent {
     }
   }
 
-  @Override
-  public RbelContentTreeNode findMessage() {
-    RbelContentTreeNode position = this;
-    while (position.getParentNode() != null) {
-      position = position.getParentNode();
-    }
-    return position;
-  }
-
-  @Override
-  public <T extends RbelFacet> boolean hasFacet(Class<T> clazz) {
-    return false;
-  }
-
   public Optional<RbelContentTreeNode> childNode(String nodeKey) {
     return Optional.ofNullable(childNodes.get(nodeKey));
   }
 
-  @Override
   public Charset getElementCharset() {
     return Optional.ofNullable(charset).orElse(StandardCharsets.UTF_8);
   }