+++ tiger-zion/src/main/java/de/gematik/test/tiger/zion/config/TigerMockResponseDescription.java
@@ -17,7 +17,7 @@ import lombok.NoArgsConstructor;
 @JsonInclude(Include.NON_NULL)
 public class TigerMockResponseDescription {
 
-  private String body;
+  @TigerSkipEvaluation private String body;
   private String bodyFile;
   @Builder.Default private Map<String, String> headers = new HashMap<>();
   @Builder.Default private Integer statusCode = 200;